put stuff in the right file
