+++ src/Clause.ts
@@ -4,7 +4,7 @@ import type Spec from './Spec';
 import type { ClauseBiblioEntry } from './Biblio';
 import type { Context } from './Context';
 
-import { getLocation } from './lint/utils';
+import { getLocation } from './utils';
 import Builder from './Builder';
 
 /*@internal*/
